requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
python-dotenv>=1.0.0
tensorflow>=2.15.0
shap>=0.44.0
reportlab>=4.1.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
